Further cleaned up Lua version
Previous commit was dumb and gay and has been replaced with an even more readable solution in which the variable `c` is scoped solely to the loop. As such, the values are shifted last, allowing the overflow to be checked on the variable `b`, which is in scope when the `until` call happens (unlike `c` now).

Overflow is simply checked via seeing if the addition result is negative. Kinda weak solution but it works flawlessly so it stays.